Store mHistoricTimes as longs intead of floats
The loss of precision was causing us to miscalculate the age of
events, and thus not detecting any pause due to denominator = 0.

Test: have a device that hasn't been rebooted for a certain
amount of time such that SystemClock#uptimeMillis can't be
accurately converted to a float, then try to swipe up and hold

Fixes: 160568387
(cherry picked from commit a14567096922f23d6e03c84fc930e3451e7471c2)
(cherry picked from commit 31ef24c9342b8d7b09f4c2f655dc2517b11d578f)

+++ quickstep/src/com/android/quickstep/util/MotionPauseDetector.java
@@ -364,7 +364,7 @@ public class MotionPauseDetector {
                         if (endPos < 0) {
                             endPos += HISTORY_SIZE;
                         }
-                        float denominator = eventTime - mHistoricTimes[endPos];
+                        long denominator = eventTime - mHistoricTimes[endPos];
                         if (denominator != 0) {
                             return (mHistoricPos[pointPos] - mHistoricPos[endPos]) / denominator;
                         }
